LayoutTests/editing/pasteboard/copy-paste-text-in-h1.html

<!DOCTYPE html>
<html>
<head>
<style>
h1 { border: solid 1px black; padding: 5px; }
</style>
</head>
<body>
<p id="description">This tests copying and pasting text in h1. WebKit should not nest h1s.
To manually test, copy and paste "hello world". You should see single border around "hello world".</p>
<div id="test" contenteditable>
<h1>hello <em>world</em></h1>
</div>
<script src="../../resources/dump-as-markup.js"></script>
<script>

var test = document.getElementById('test');
test.focus();

Markup.description(document.getElementById('description').textContent);

document.execCommand('SelectAll', false, null);
document.execCommand('Copy', false, null);
document.execCommand('Paste', false, null);

if (window.layoutTestController)
    Markup.dump(test);
else
    Markup.noAutoDump();

</script>
</body>
</html>
