LayoutTests/editing/pasteboard/copy-paste-text-in-h1-expected.txt

This tests copying and pasting text in h1. WebKit should not nest h1s.
To manually test, copy and paste "hello world". You should see single border around "hello world".
| "
"
| <h1>
|   "hello"
|   " "
|   <em>
|     "world<#selection-caret>"
| "
"
